Fixed problem with topic removal due to metrics
There is validation on topic removal that there are no subscriptions on
that topic. But in the same time the metric which has subpath same as
subscription path is stored (e.g. subscription path is
/groups/{group}/topics/{topic}/subscriptions/{sub} and metric path is
/groups/{group}/topics/{topic}/subscriptions/{sub}/metrics/delivered
Validation checks only if path exists.

The situation when there were no subscription but metric was stored
may be caused by missing subscription check while storing inflight
counter or because of the race condition (removing topic and storing
metric at the same time).

The solution is to check if data for subscription path is actually
present and also to add missing check.