Flag 'fallback to remote datacenter'
This is a preliminary step to introduce fallback to a remote datacenter in case of any failures with the local Kafka cluster while publishing. The flag introduced here, when set to true, will enable the fallback for a given topic and disable persistent buffers for it. The mechanism respecting this flag will be added as a follow-up.